♻️ Improve: integrate Project data formatting functions [#144]
